tensorflow recon now cleaned up
